+++ Drop-In/src/main/java/com/braintreepayments/api/DropInClient.java
@@ -167,9 +167,6 @@ public class DropInClient {
     }
 
     void tokenizeVenmoAccount(FragmentActivity activity, VenmoTokenizeAccountCallback callback) {
-        // TODO: Add VenmoRequest setter to DropInRequest and remove DropInRequest#shouldVaultVenmo()
-//        VenmoRequest venmoRequest = new VenmoRequest(VenmoPaymentMethodUsage.SINGLE_USE);
-//        venmoRequest.setShouldVault(dropInRequest.getVaultVenmoDefaultValue());
         venmoClient.tokenizeVenmoAccount(activity, dropInRequest.getVenmoRequest(), callback);
     }