+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Braintree Android Drop-In Release Notes
 
 ## unreleased
+* Venmo
+  * Add `VenmoRequest` setter to `DropInRequest`
 * Braintree Core
   * Bump braintree_android module dependency versions to `4.7.0`
 * CardForm
@@ -14,6 +16,8 @@
     * Make `PaymentMethodItemView` package-private
     * Make `SupportedPaymentMethodsViewHolder` package-private
     * Make `VaultedPaymentMethodSelectedListener` package-private
+  * Venmo
+    * Remove `vaultVenmoDefaultValue` in favor of setting `shouldVault` on `VenmoRequest`
 
 ## 6.0.0-beta2